requirements.txt
numpy~=1.22.4
torch~=1.13.1
tqdm~=4.64.1
wandb~=0.15.0
torchvision~=0.14.1
Pillow~=9.1.1